requirements.txt
numpy
scipy
resampy
tensorflow
six
pysoundfile
ffmpeg-python
pydub
matplotlib
keras
flask
flask-cors
flask-restful